+++ NEWS.md
@@ -1,5 +1,6 @@
 # nlmixr2est (development version)
 
+
 - Add `$fitMergeFull`, `$fitMergInner`, `$fitMergeLeft`,
   `$fitMergeRight` as a complement to `$dataMergeFull`,
   `$dataMergInner`, `$dataMergeLeft`, `$dataMergeRight`.  The fit
@@ -8,7 +9,11 @@
   since the `DV` in the fit simulates values under the ipred/residual
   assumption and will give more appropriate goodness of fits,
   otherwise these values are the limit of whatever censoring is
-  applied.
+  applied
+  
+- Moved the mu reference fix for the split mu referenced model here
+  (from babelmixr2)
+
 
 # nlmixr2est 2.1.4